Move tests related to Clock into their own module
