+++ tests/test_cases/test_cocotb/Makefile
@@ -39,6 +39,7 @@ MODULE := "\
 	test_generator_coroutines,\
 	test_timing_triggers,\
 	test_scheduler,\
+	test_clock,\
 	"
 
 ifeq ($(shell python -c "import sys; print(sys.version_info >= (3, 6))"), "True")
